server-plano/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
google-generativeai
supabase
python-jose[cryptography]
passlib[bcrypt]
redis
hiredis
